Reorganize src tree into some folders
